+++ iModelCore/ECDb/CHANGES.md
@@ -5,7 +5,12 @@ This document including important changes to syntax or file format.
 | Module  | Version   |
 | ------- | --------- |
 | Profile | `4.0.0.5` |
-| ECSQL   | `2.0.3.2` |
+| ECSQL   | `2.0.3.3` |
+
+## ## `06/24/2026`: Added PRAGMAs `ecdb_known_features` and `ecdb_used_features`
+* ECSql version change `2.0.3.2` -> `2.0.3.3`.
+* Added `PRAGMA ecdb_known_features` which returns a list of the features supported by the ECDb runtime.
+* Added `PRAGMA ecdb_used_features` which returns a list of features that are being used in the ECDb file.
 
 ## ## `04/24/2026`: Allow optional ON clause with CROSS JOIN
 * ECSql version change `2.0.3.1` -> `2.0.3.2`.